fix: hmr linting issue
