+++ services/design-system/src/SelectChangeEvent.ts
@@ -1,0 +1,8 @@
+export type SelectChangeEvent<T = string> =
+  | (Event & {
+      target: {
+        value: T;
+        name: string;
+      };
+    })
+  | React.ChangeEvent<HTMLInputElement>;
