+++ services/design-system/src/components/Alert/Alert.tsx
@@ -1,6 +1,6 @@
 import MUIAlert, { AlertProps as MUIAlertProps } from '@mui/material/Alert';
 
-export type AlertProps = {} & MUIAlertProps;
+export type AlertProps = MUIAlertProps;
 
 export const Alert = (props: AlertProps) => {
   const { ...muiProps } = props;
