+++ services/design-system/src/components/Button/LoadingButton.tsx
@@ -2,7 +2,9 @@ import MuiLoadingButton, {
   LoadingButtonProps as MuiLoadingButtonProps,
 } from '@mui/lab/LoadingButton';
 
-export type LoadingButtonProps = {} & MuiLoadingButtonProps;
+export type LoadingButtonProps = {
+  selected?: boolean;
+} & MuiLoadingButtonProps;
 
 export const LoadingButton = (props: LoadingButtonProps) => {
   const { children, ...muiProps } = props;
@@ -10,8 +12,11 @@ export const LoadingButton = (props: LoadingButtonProps) => {
     <MuiLoadingButton
       {...muiProps}
       style={{
+        ...(props.selected && { backgroundColor: '#616366', color: '#FFF' }),
         fontFamily: 'Helvetica, arial, sans-serif',
         fontWeight: 'bold',
+        borderRadius: '36px',
+        padding: '15px 30px 10px 30px',
       }}
     >
       {children}
