+++ services/design-system/src/components/Checkbox/Checkbox.tsx
@@ -2,7 +2,7 @@ import MUICheckbox, {
   CheckboxProps as MUICheckboxProps,
 } from '@mui/material/Checkbox';
 
-export type CheckboxProps = {} & MUICheckboxProps;
+export type CheckboxProps = MUICheckboxProps;
 
 export const Checkbox = (props: CheckboxProps) => {
   const { ...muiProps } = props;
