+++ services/design-system/src/components/CodeSnippet/CodeSnippetProps.ts
@@ -1,0 +1,11 @@
+export enum CodeSnippetLanguages {
+  JSON = 'json',
+  HTML = 'html',
+  CSS = 'css',
+  JAVASCRIPT = 'javascript',
+}
+
+export type CodeSnippetProps = {
+  data: string;
+  language: CodeSnippetLanguages;
+};
