+++ services/design-system/src/components/Gauge/GaugeProps.ts
@@ -1,0 +1,15 @@
+export enum OrientationType {
+  RIGHT = 'RIGHT',
+  LEFT = 'LEFT',
+}
+
+export type GaugeProps = {
+  width: number;
+  height: number;
+  color?: string;
+  strokeWidth?: number;
+  orientation?: OrientationType;
+  current?: number;
+  max?: number;
+  boundingBox: boolean;
+};
