+++ services/design-system/src/components/Link/Link.tsx
@@ -1,6 +1,6 @@
 import MUILink, { LinkProps as MUILinkProps } from '@mui/material/Link';
 
-export type LinkProps = {} & MUILinkProps;
+export type LinkProps = MUILinkProps;
 
 export const Link = (props: LinkProps) => {
   const { children, ...muiProps } = props;
