--- services/design-system/src/components/Mui/index.tsx
@@ -1,46 +1,0 @@
-/* eslint-disable import/extensions */
-
-export * from './Button';
-export * from './Checkbox';
-export * from './Spinner';
-export * from './TextField';
-export * from './Typography';
-export * from './LoadingButton';
-export * from './Link';
-export * from './Layout';
-export * from './Alert';
-
-export type SelectChangeEvent<T = string> =
-  | (Event & {
-    target: {
-      value: T;
-      name: string;
-    };
-  })
-  | React.ChangeEvent<HTMLInputElement>;
-
-export {
-  Drawer,
-  CssBaseline,
-  Toolbar,
-  List,
-  Divider,
-  IconButton,
-  ListItem,
-  ListItemButton,
-  FormControl,
-  InputLabel,
-  Select,
-  Box,
-  CardContent,
-  Card,
-  Grid,
-  Stack,
-  Container,
-  OutlinedInput,
-  MenuItem,
-  ListItemIcon,
-  ListItemText,
-  styled,
-  useTheme,
-} from '@mui/material';
