+++ services/design-system/src/components/Typography/Typography.tsx
@@ -1,0 +1,8 @@
+import { Typography as MuiTypography, TypographyProps as MuiTypographyProps } from '@mui/material';
+import './style.module.scss'
+
+export type TypographyProps = MuiTypographyProps;
+
+export const Typography = MuiTypography;
+
+export default Typography;
