+++ services/design-system/src/main.ts
@@ -7,5 +7,15 @@ export * from './components/FileUpload/FileUpload';
 export * from './components/Gauge/Gauge';
 export * from './components/Meter/Meter';
 export * from './components/PageFooter/PageFooter';
-export * from './components/NonInteractiveLink/NonInteractiveLink';
+export * from './components/Link/NonInteractiveLink';
 export * from './components/ColorPallette';
+export * from './components/Typography/Typography';
+export * from './components/Alert/Alert';
+export * from './components/TextField/TextField';
+export * from './components/Button/Button';
+export * from './components/Checkbox/Checkbox';
+export * from './components/Spinner/Spinner';
+export * from './components/Button/LoadingButton';
+export * from './components/Link/Link';
+export * from './components/Layout/Layout';
+export * from './SelectChangeEvent';
